pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Giovani Liskoski Zanini",email = "giovanilzanini@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.9.7 || >3.9.7,<3.11"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "plotly-express (>=0.4.1,<0.5.0)",
    "streamlit (>=1.45.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

